backend_realtime/requirements.txt
flask==3.0.0
flask-socketio==5.3.5
eventlet==0.33.3
pymongo==4.6.1
python-dotenv==1.0.0
flask-cors==4.0.0

